Unpolished version 1.0 bugfixes

+++ Projekt/ViewModels/FileViewModel.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
+using System.Threading;
 using System.Windows;
 using System.Windows.Input;
 using Projekt.Commands;
@@ -9,17 +11,17 @@ namespace Projekt
 {
     public class FileViewModel : BasicViewModel
     {
-
         public FileViewModel()
         {
             DeleteCommand = new RelayCommand(Delete);
             ShowDataCommand = new RelayCommand(ShowData);
+            OpenDirectoryCommand = new RelayCommand(OpenDirectory);
         }
-
 
         #region Public properties
 
         private string _fileName = "";
+
         public string FileName
         {
             get => _fileName;
@@ -28,7 +30,6 @@ namespace Projekt
                 if (_fileName != value)
                 {
                     _fileName = value;
-                    new Serializer().UpdateXmlFile(FilePath, "GivenName", FileName);
                     SetDisplayDataFrameTitle(FileName);
                 }
             }
@@ -40,8 +41,11 @@ namespace Projekt
 
         public FilesListViewModel Parent { get; set; }
 
+        public bool WasEvaluated { get; set; }
+
         public PresentationScreenViewModel DisplayDataFrame { get; set; }
 
+        public FileControlRepresents Represents { get; set; }
 
         #endregion
 
@@ -51,30 +55,41 @@ namespace Projekt
 
         public ICommand ShowDataCommand { get; set; }
 
+        public ICommand OpenDirectoryCommand { get; set; }
+
         #endregion
 
         #region Command methods
+
         private void Delete(object obj)
         {
-            Parent?.List.Remove(this);
-            File.Delete(FilePath);
+            if (Parent != null)
+            {
+                Parent.List.Remove(this);
+                File.Delete(FilePath);
+            }
         }
 
         private void ShowData(object obj)
         {
+            //TODO: add show data for DTMC variant, for now return
+
+            if (Represents == FileControlRepresents.DTMC) return;
+            if (!WasEvaluated) return;
+
             if (DisplayDataFrame != null)
             {
-                Serializer serializer = new Serializer();
+                var serializer = new Serializer();
 
                 try
                 {
-                    ResultsViewModel viewmodel = serializer.ReadFromXmlFile<ResultsViewModel>(FilePath);
-
-                    Results dataFrame = new Results(viewmodel);
+                    var dataFrame = serializer.ReadFromJsonFile<TestedGroup>(FilePath).AsResultsControl();
 
                     DisplayDataFrame.ContentPresented = dataFrame;
 
                     DisplayDataFrame.Title = FileName;
+
+                    DisplayDataFrame.HasResultsShown = true;
                 }
                 catch (Exception ex)
                 {
@@ -83,10 +98,19 @@ namespace Projekt
             }
         }
 
+        private void OpenDirectory(object obj)
+        {
+            if (File.Exists(FilePath)) Process.Start(FilePath);
+        }
+        #endregion
+
+        #region Private mathods
+
         private void SetDisplayDataFrameTitle(string title)
         {
             if (DisplayDataFrame != null) DisplayDataFrame.Title = title;
         }
+
         #endregion
     }
 }
